+++ test/filter-generator/TestRegularScriptGenerator.cpp
@@ -51,7 +51,7 @@ BOOST_AUTO_TEST_CASE(should_generate_ffmpeg_script)
     "delogo=enable='between(n,0,499)':x=10:y=11:w=12:h=13,\n"
     "drawbox=enable='between(n,500,999)':x=20:y=21:w=22:h=23:c=black:t=max,\n"
     "drawbox=enable='between(n,1300,1999)':x=30:y=31:w=32:h=33:c=black:t=max,\n"
-    "drawbox=enable='gte(n,2000)':x=40:y=41:w=42:h=43:c=black:t=max\n";
+    "drawbox=enable='gte(n,2000)':x=40:y=41:w=42:h=43:c=black:t=max";
   BOOST_CHECK_EQUAL(out.str(), expected);
 }
 
@@ -67,6 +67,21 @@ BOOST_AUTO_TEST_CASE(should_discard_a_null_filter_at_the_end)
   g->generate_ffmpeg_script(out);
 
   std::string expected =
-    "delogo=enable='between(n,0,999)':x=10:y=11:w=12:h=13,\n";
+    "delogo=enable='between(n,0,999)':x=10:y=11:w=12:h=13";
+  BOOST_CHECK_EQUAL(out.str(), expected);
+}
+
+
+BOOST_AUTO_TEST_CASE(should_work_for_a_one_filter_list)
+{
+  FilterList list;
+  list.insert(50, new DelogoFilter(10, 11, 12, 13));
+  std::shared_ptr<ScriptGenerator> g = RegularScriptGenerator::create(list);
+
+  std::ostringstream out;
+  g->generate_ffmpeg_script(out);
+
+  std::string expected =
+    "delogo=enable='gte(n,49)':x=10:y=11:w=12:h=13";
   BOOST_CHECK_EQUAL(out.str(), expected);
 }